requirements.txt
flask
sqlalchemy
flask-marshmallow
Flask-RESTful
marshmallow-sqlalchemy
python-dotenv
pymysql
Flask-JWT-Extended